# docs/swagger.yaml
definitions:
  models.AliasListResponse:
    properties:
      aliases:
        items:
          $ref: '#/definitions/models.AliasResponse'
        type: array
    type: object
  models.AliasResponse:
    properties:
      alias:
        type: string
      email:
        type: string
    type: object
  models.EmailListResponse:
    properties:
      emails:
        items:
          type: string
        type: array
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.StatusResponse:
    properties:
      running:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: admin@scheid.tech
    name: Christian Scheid
    url: https://github.com/scheidti/docker-mailserver-aliases
  description: API for managing aliases in a Docker Mailserver container
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Docker Mailserver Aliases API
  version: "1.0"
paths:
  /v1/aliases:
    get:
      consumes:
      - application/json
      description: Gets a list of all available email aliases from the Docker Mailserver
        container
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AliasListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List of all available email aliases
      tags:
      - Aliases
    post:
      consumes:
      - application/json
      description: Adds a new email alias to the Docker Mailserver container
      parameters:
      - description: Alias to add
        in: body
        name: alias
        required: true
        schema:
          $ref: '#/definitions/models.AliasResponse'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AliasResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add a new email alias
      tags:
      - Aliases
  /v1/aliases/{alias}:
    delete:
      consumes:
      - application/json
      description: Deletes an email alias from the Docker Mailserver container
      parameters:
      - description: Alias to delete
        in: path
        name: alias
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete an email alias
      tags:
      - Aliases
  /v1/emails:
    get:
      consumes:
      - application/json
      description: Gets a list of all available email addresses from the Docker Mailserver
        container
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EmailListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List of all available email addresses
      tags:
      - E-Mails
  /v1/status:
    get:
      consumes:
      - application/json
      description: Checks if the Docker Mailserver Docker container is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.StatusResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Checks Mailserver Docker container
      tags:
      - Utility
swagger: "2.0"